JL8-24 refactor: 마이페이지 playlist부분 리팩토링링
