+++ src/app/my-page/_components/Playlist.tsx
@@ -1,85 +1,31 @@
 'use client'
-import { updatePlaylistLike } from '@/api/my-page/actions'
 import { PlaylistUI } from '@/components/common'
+import { usePlaylistQuery } from '@/hooks/usePlaylistQuery'
+import { useToggleLikeMutation } from '@/hooks/useToggleLikeMutation'
 import { userStore } from '@/store/userSlice'
-import type { OldData } from '@/types/playlistLike'
-import {
-  useInfiniteQuery,
-  useMutation,
-  useQueryClient,
-} from '@tanstack/react-query'
 import { useRouter } from 'next/navigation'
 import { useInView } from 'react-intersection-observer'
-import { fetchUserPlayList } from './fetchUserPlayList'
 import MyPageSkeleton from './MyPageSkeleton'
 
 const PlayList = () => {
   const { user } = userStore()
   const router = useRouter()
-  const queryClient = useQueryClient()
   const {
-    data: playlists,
+    playlists,
     fetchNextPage,
     hasNextPage,
     isFetchingNextPage,
     error,
     isLoading,
-  } = useInfiniteQuery({
-    queryKey: ['playlist'],
-    queryFn: ({ pageParam = 0 }) => fetchUserPlayList({ pageParam }),
-    getNextPageParam: (lastPage) => lastPage?.nextCursor || undefined,
-    getPreviousPageParam: (firstPage) => firstPage?.prevCursor || undefined,
-    initialPageParam: 0,
-  })
+  } = usePlaylistQuery()
+
+  const toggleLike = useToggleLikeMutation()
   const { ref } = useInView({
     threshold: 1,
     onChange: (inView) => {
       if (inView && hasNextPage && !isFetchingNextPage) {
         fetchNextPage()
       }
-    },
-  })
-  const toggleLike = useMutation({
-    mutationFn: updatePlaylistLike,
-    onMutate: async (newLikeData: { playlist_id: string; user_id: string }) => {
-      await queryClient.cancelQueries({ queryKey: ['playlist'] })
-      const prevPlaylists = queryClient.getQueryData<OldData>(['playlist'])
-      queryClient.setQueryData(['playlist'], (prevData: OldData) => {
-        if (!prevData) return prevData
-        return {
-          ...prevData,
-          pages: prevData.pages.map((page) => ({
-            ...page,
-            playlists: page.playlists.map((item) => {
-              if (item.id === newLikeData.playlist_id) {
-                return {
-                  ...item,
-                  playlist_like: item.playlist_like.some(
-                    (like) => like.user_id === newLikeData.user_id,
-                  )
-                    ? item.playlist_like.filter(
-                        (like) => like.user_id !== newLikeData.user_id,
-                      ) // 좋아요 취소
-                    : [...item.playlist_like, { user_id: newLikeData.user_id }],
-                }
-              }
-              return item
-            }),
-          })),
-        }
-      })
-      return { prevPlaylists }
-    },
-    onError: (err, newLikeData, context) => {
-      if (context?.prevPlaylists) {
-        queryClient.setQueryData(['playlist'], context.prevPlaylists)
-      }
-    },
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ['playlist'] })
-    },
-    onSettled: () => {
-      queryClient.invalidateQueries({ queryKey: ['playlist'] })
     },
   })
   if (isLoading)
@@ -99,7 +45,6 @@ const PlayList = () => {
                   (like) => like.user_id === p.user_id,
                 )
                 const likeCount = p.playlist_like.length
-                // 데이터리소스가 많이 낭비됨
                 return (
                   <PlaylistUI
                     key={p.id}
