+++ src/app/my-page/_components/fetchUserPlayList.ts
@@ -1,3 +1,4 @@
+import { userStore } from '@/store/userSlice'
 import { supabase } from '@/utils/supabase/client'
 
 type Param = {
@@ -5,11 +6,11 @@ type Param = {
 }
 
 export const fetchUserPlayList = async ({ pageParam = 0 }: Param) => {
-  const { data } = await supabase.auth.getUser()
+  const { user } = userStore.getState()
   const { data: playlists, error } = await supabase
     .from('playlists')
     .select(`*,playlist_like!left(user_id)`)
-    .eq('user_id', data.user!.id)
+    .eq('user_id', user?.id!)
     .range(pageParam * 10, (pageParam + 1) * 10 - 1)
   if (error) {
     console.error(error.message)
